move deepmd.entrypoints.{doc,gui} to deepmd_utils.entrypoints.{doc,gui}
